# swagger.yaml
openapi: 3.0.0
servers:
  - url: 'http://localhost:8080/'
    description: An description
info:
  description: APIs Documentation Swagger - Entrega Final ddelcampo
  version: 1.0.1
  title: Entrega Final ddelcampo - Curso Backend CoderHouse
  contact:
    email: ddelcampo@teco.com.ar
...
paths:
...
  /api/productos:
    get:
      summary: 'http://localhost:8080/api/productos?secret_token=<TOKEN>'
      parameters:
        - in: query
          name: secret_token
          schema:
            type: string
            example: <TOKEN>
      tags:
        - Default
      responses:
        '200':
          description: 'http://localhost:8080/api/productos?secret_token=<TOKEN>'
          content:
            application/json:
              schema:
                example:
                  - _id: 624388d2fcd77b6221503fac
                    nombre: Mochila
                    precio: 1235
                    descripcion: mochi
                    categoria: mochi
                    stock: 1224
                    foto: https://cdn3.iconfinder.com/data/icons/education-209/64/bag-pack-container-school-128.png
                    timestamp: '29/3/2022 19:31:55'
                    id: 1
                    __v: 0
                  - _id: 624446394c4152bd8dbcea43
                    nombre: Calculadora
                    precio: 123
                    descripcion: Calculadora Cientifica
                    categoria: Calculadoras
                    stock: 1119
                    foto: https://cdn3.iconfinder.com/data/icons/education-209/64/calculator-math-tool-school-256.png
                    timestamp: '30/3/2022 08:59:53'
                    id: 2
                    __v: 0
                  - _id: 6244a27927996b93d2f0b015
                    nombre: Paleta
                    precio: 12900
                    descripcion: Paleta de colores
                    categoria: Paleta
                    stock: 1
                    foto: https://cdn3.iconfinder.com/data/icons/education-209/64/paint-color-pallete-brush-academy-128.png
                    timestamp: '30/3/2022 15:33:39'
                    id: 3
                    __v: 0
    post:
      summary: 'http://localhost:8080/api/productos?secret_token=<TOKEN>'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                nombre: Pizarra
                precio: 2500
                descripcion: Pizarra 4x4
                categoria: Pizarras
                stock: 185
                foto: https://cdn3.iconfinder.com/data/icons/education-209/64/board-math-class-school-128.png
      parameters:
        - in: query
          name: secret_token
          schema:
            type: string
            example: <TOKEN>
      tags:
        - Default
      responses:
        '200':
          description: 'http://localhost:8080/api/productos?secret_token=<TOKEN>'
          content:
            application/json:
              schema:
                example: 'Nuevo producto ID: 4 cargado OK!'
...
  /api/carrito/<ID>/productos:
    get:
      summary: 'http://localhost:8080/api/carrito/1/productos?secret_token=<TOKEN>'
      parameters:
        - in: query
          name: secret_token
          schema:
            type: string
            example: <TOKEN>
      tags:
        - Default
      responses:
        '200':
          description: 'http://localhost:8080/api/carrito/<ID>/productos?secret_token=<TOKEN>'
          content:
            application/json:
              schema:
                example:
                  - nombre: Calculadora 2
                    precio: 2500
                    descripcion: Calculadora Cientifica 2
                    foto: https://cdn3.iconfinder.com/data/icons/education-209/64/calculator-math-tool-school-128.png
                    cantidad: 2
                    timestamp: '10/3/2022 11:46:00'
                    id: 2
                  - nombre: Calculadora
                    precio: 2500
                    descripcion: Calculadora Cientifica
                    foto: https://cdn3.iconfinder.com/data/icons/education-209/64/calculator-math-tool-school-128.png
                    cantidad: 3
                    timestamp: '10/3/2022 11:45:47'
                    id: 1
  /api/carrito:
    post:
      summary: 'http://localhost:8080/api/carrito?secret_token=<TOKEN>'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: example@example.com
                direccion: Av. Siempre Viva 123
      parameters:
        - in: query
          name: secret_token
          schema:
            type: string
            example: <TOKEN>
      tags:
        - Default
      responses:
        '200':
          description: 'http://localhost:8080/api/carrito?secret_token=<TOKEN>'
          content:
            application/json:
              schema:
                example: 76
...
  /api/usuario/register:
    post:
      summary: 'http://localhost:8080/api/usuario/register'
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: example1@example.com
                password:
                  type: string
                  example: password
                passwordConfirm:
                  type: string
                  example: password
                nombre:
                  type: string
                  example: Nombre Apellido
                telefono:
                  type: string
                  example: '1123456789'
                direccion:
                  type: string
                  example: Av Siempre Viva
      tags:
        - Default
      responses:
        '200':
          description: 'http://localhost:8080/signup'
          content:
            application/json:
              schema:
                example:
                  message: El usuario se registro con exito
                  userID:
                    userId: 1
  /api/usuario/login:
    post:
      summary: 'http://localhost:8080/api/usuario/login'
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: example@example
                password:
                  type: string
                  example: password
      tags:
        - Default
      responses:
        '200':
          description: 'http://localhost:8080/login'
          content:
            application/json:
              schema:
                example:
                  token: <TOKEN>
...
components:
  securitySchemes: {}